integration-tests: split workflow tests out
