+++ integration-tests/cli/jobs/wf-count.json
@@ -2,10 +2,12 @@
   "start": "a",
   "jobs": {
     "a": {
+      "adaptor": "common",
       "expression": "fn((state) => (state.data.count ? state : { data: { count: 21 } }));",
       "next": { "b": true }
     },
     "b": {
+      "adaptor": "common",
       "expression": "fn((state) => { state.data.count = state.data.count * 2; return state; });"
     }
   }
